refactor: index 내 html와 상호작용 기능 분리
- mainPage와 resultPage로 분리
- controller에서 두 파일을 제어
